policyfilter_test: use ResetStateOnlyForTesting()
In my development testing, I'm running the TestNamespacedPolicies test
directly via -test.run or similar. For CI, however, other tests might
run before it that call GetState(). This means that even if we set
option.Config.EnablePolicyFilter to true, it will not be effective.

Add a ResetStateOnlyForTesting that allows us to test policyfilter by
resetting the state.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>